\name{aov_treat_compare}
\alias{aov_treat_compare}
\title{
  ANOVA test across treeatment conditions
}
\description{

}
\usage{
  aov_treat_compare(dat, var_names, treat_var)
}

\arguments{
  \item{dat}{

  }
  \item{var_names}{

  }
  \item{treat_var}{

  }
}
\value{

}
\references{

}
\author{
  Loren Collingwood <loren.collingwood@ucr.edu>
}

\examples{
  }
