\name{multiplot}
\alias{multiplot}
\title{
  Multiplot
}
\description{Makes quadrant of ggplots

}
\usage{
  multiplot(..., plotlist=NULL, file, cols=1, layout=NULL)
}

\arguments{
  \item{...}{saved plot objects

  }
  \item{plotlist}{NULL

  }
  \item{file}{path to save file

  }
  \item{cols}{Number of columns, 1 or 2

  }
  \item{layout}{NULL

  }

}
\value{New plot

}
\references{

}
\author{
  Loren Collingwood <loren.collingwood@ucr.edu>
}

\examples{
}
