\name{Rmturkcheck-package.Rd}
\alias{Rmturkcheck-package.Rd}
\alias{Rmturkcheck}
\docType{package}
\title{
  Evaluates data post wave 1 wave 2 MTurk experiments/surveys
}
\description{Contains mturk clean id functions, balance f-tests across experimental groups, t-test functions for attrition comparison, and functions for summarizing variables.
  
}
%\details{
%See demo(demo, "Rvoterdistance") for examples on how to use the code}
\author{
Loren Collingwood, UC Riverside

Maintainer: Loren Collingwood <loren.collingwood@ucr.edu>
}
\references{
  Collingwood, Lajevardi, Oskooii (2018)
}
%\keyword{ package }
%\seealso{
%  Optional links to other man pages
%}
%\examples{
%  \dontrun{
%     ## Optional simple examples of the most important functions
%     ## These can be in \dontrun{} and \donttest{} blocks.   
%  }
%}
