\name{sum_func}
\alias{sum_func}
\title{
  sum_func
}
\description{Summary of descriptive statistics; usually done for paper appendix

}
\usage{
  sum_func(dat)
}

\arguments{
  \item{dat}{Dataset of class dataframe with variables to describe
  }

}
\value{Table/data.frame of variables as rows, column names: Minimum, Maximum, Mean, Median, Standard Deviation

}
\references{

}
\author{
  Loren Collingwood <loren.collingwood@ucr.edu>
}

\examples{
  data(w2)
  sum_func(w2)
}
