\name{multiplot}
\alias{multiplot}
\title{
  multiplot
}
\description{Makes plots

}
\usage{
  multiplot(..., plotlist=NULL, file, cols=1, layout=NULL)
}

\arguments{
  \item{...}{

  }
  \item{plotlist}{

  }
  \item{file}{

  }
  \item{cols}{

  }
  \item{layout}{

  }

}
\value{

}
\references{

}
\author{
  Loren Collingwood <loren.collingwood@ucr.edu>
}

\examples{
}
