\name{sim_extract}
\alias{sim_extract}
\title{
  sim_extract
}
\description{Extract simulations from Zelig model

}
\usage{
  sim_extract(sim_out, conf_band=F)
}

\arguments{
  \item{sim_out}{

  }
  \item{conf_band}{

  }

}
\value{

}
\references{

}
\author{
  Loren Collingwood <loren.collingwood@ucr.edu>
}

\examples{
}
