\name{t_generalize}
\alias{t_generalize}
\title{
  t_generalize
}
\description{Ttests across variables producing table output

}
\usage{
  t_generalize(w1, w2, cnames1, cnames2)
}

\arguments{
  \item{w1}{

  }
  \item{w2}{

  }
  \item{cnames1}{

  }
  \item{cnames2}{

  }

}
\value{

}
\references{

}
\author{
  Loren Collingwood <loren.collingwood@ucr.edu>
}

\examples{
}
