\name{t_out}
\alias{t_out}
\title{
  t_out
}
\description{Single t-test comparison producing table

}
\usage{
  t_out(test)
}

\arguments{
  \item{test}{t-test output between two vectors
  }

}
\value{Table of values

}
\references{

}
\author{
  Loren Collingwood <loren.collingwood@ucr.edu>
}

\examples{
}
